Delete `IdentityGadget` and replace with `BasePrecompileGadget`.
