+++ zkevm-circuits/src/evm_circuit/execution/precompiles/mod.rs
@@ -12,9 +12,6 @@ use crate::{
     witness::{Block, Call, ExecStep, Transaction},
 };
 
-mod identity;
-pub use identity::IdentityGadget;
-
 #[derive(Clone, Debug)]
 pub struct BasePrecompileGadget<F, const S: ExecutionState> {
     is_success: Cell<F>,
@@ -104,5 +101,113 @@ impl<F: Field, const S: ExecutionState> ExecutionGadget<F> for BasePrecompileGad
         )?;
 
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use bus_mapping::{
+        evm::{OpcodeId, PrecompileCallArgs},
+        precompile::PrecompileCalls,
+    };
+    use eth_types::{bytecode, word, ToWord};
+    use itertools::Itertools;
+    use mock::TestContext;
+
+    use crate::test_util::CircuitTestBuilder;
+
+    lazy_static::lazy_static! {
+        static ref TEST_VECTOR: Vec<PrecompileCallArgs> = {
+            vec![
+                PrecompileCallArgs {
+                    name: "single-byte success",
+                    setup_code: bytecode! {
+                        // place params in memory
+                        PUSH1(0xff)
+                        PUSH1(0x00)
+                        MSTORE
+                    },
+                    call_data_offset: 0x1f.into(),
+                    call_data_length: 0x01.into(),
+                    ret_offset: 0x3f.into(),
+                    ret_size: 0x01.into(),
+                    address: PrecompileCalls::Identity.address().to_word(),
+                    ..Default::default()
+                },
+                PrecompileCallArgs {
+                    name: "multi-bytes success (less than 32 bytes)",
+                    setup_code: bytecode! {
+                        // place params in memory
+                        PUSH16(word!("0x0123456789abcdef0f1e2d3c4b5a6978"))
+                        PUSH1(0x00)
+                        MSTORE
+                    },
+                    call_data_offset: 0x00.into(),
+                    call_data_length: 0x10.into(),
+                    ret_offset: 0x20.into(),
+                    ret_size: 0x10.into(),
+                    address: PrecompileCalls::Identity.address().to_word(),
+                    ..Default::default()
+                },
+                PrecompileCallArgs {
+                    name: "multi-bytes success (more than 32 bytes)",
+                    setup_code: bytecode! {
+                        // place params in memory
+                        PUSH30(word!("0x0123456789abcdef0f1e2d3c4b5a6978"))
+                        PUSH1(0x00) // place from 0x00 in memory
+                        MSTORE
+                        PUSH30(word!("0xaabbccdd001122331039abcdefefef84"))
+                        PUSH1(0x20) // place from 0x20 in memory
+                        MSTORE
+                    },
+                    // copy 63 bytes from memory addr 0
+                    call_data_offset: 0x00.into(),
+                    call_data_length: 0x3f.into(),
+                    // return only 35 bytes and write from memory addr 72
+                    ret_offset: 0x48.into(),
+                    ret_size: 0x23.into(),
+                    address: PrecompileCalls::Identity.address().to_word(),
+                    ..Default::default()
+                },
+                /* TODO(rohit): debug error cases
+                PrecompileCallArgs {
+                    name: "insufficient gas (precompile call should fail)",
+                    setup_code: bytecode! {
+                        // place params in memory
+                        PUSH16(word!("0x0123456789abcdef0f1e2d3c4b5a6978"))
+                        PUSH1(0x00)
+                        MSTORE
+                    },
+                    call_data_offset: 0x00.into(),
+                    call_data_length: 0x10.into(),
+                    ret_offset: 0x20.into(),
+                    ret_size: 0x10.into(),
+                    address: PrecompileCalls::Identity.address().to_word(),
+                    // set gas to be insufficient
+                    gas: 1.into(),
+                    ..Default::default()
+                },
+                */
+            ]
+        };
+    }
+
+    #[test]
+    fn precompile_identity_test() {
+        let call_kinds = vec![
+            OpcodeId::CALL,
+            OpcodeId::STATICCALL,
+            OpcodeId::DELEGATECALL,
+            OpcodeId::CALLCODE,
+        ];
+
+        for (test_vector, &call_kind) in TEST_VECTOR.iter().cartesian_product(&call_kinds) {
+            let bytecode = test_vector.with_call_op(call_kind);
+
+            CircuitTestBuilder::new_from_test_ctx(
+                TestContext::<2, 1>::simple_ctx_with_bytecode(bytecode).unwrap(),
+            )
+            .run();
+        }
     }
 }
